Create `Input` and `Output` structs, write tests
`Input` and `Output` represent the input and output of the `convert` command. They include the type and a reader and writer, respectively.

This enables us to abstract where input and output comes from and goes to (e.g. files, stdin/stdout, etc.)

This also enabled us to write tests, which have been included for the new `convert::run()` function which handles running the `convert` command.